[ext-msg] MessageFetchingCommander (adds aux data to msg)
