+++ kifi-backend/modules/eliza/app/com/keepit/eliza/controllers/mobile/MobileMessagingController.scala
@@ -1,7 +1,7 @@
 package com.keepit.eliza.controllers.mobile
 
 import com.keepit.social.BasicUserLikeEntity._
-import com.keepit.eliza.commanders.{MessagingUtils, NotificationCommander, MessagingCommander}
+import com.keepit.eliza.commanders._
 import com.keepit.eliza.model.MessageThread
 import com.keepit.common.controller.{ElizaServiceController, MobileController, ActionAuthenticator}
 import com.keepit.common.time._
@@ -18,12 +18,16 @@ import scala.concurrent.Future
 
 import com.google.inject.Inject
 import com.keepit.social.{BasicUserLikeEntity, BasicUser}
+import play.api.libs.json.JsArray
+import play.api.libs.json.JsString
+import scala.Some
+import play.api.libs.json.JsObject
 
 
 class MobileMessagingController @Inject() (
   messagingCommander: MessagingCommander,
+  basicMessageCommander: MessageFetchingCommander,
   notificationCommander: NotificationCommander,
-  messagingUtils: MessagingUtils,
   actionAuthenticator: ActionAuthenticator,
   heimdalContextBuilder: HeimdalContextBuilderFactory
   ) extends MobileController(actionAuthenticator) with ElizaServiceController {
@@ -81,50 +85,40 @@ class MobileMessagingController @Inject() (
 
 
   def getThread(threadId: String) = JsonAction.authenticatedAsync { request =>
-      messagingCommander.getThreadMessagesWithBasicUser(ExternalId[MessageThread](threadId), None) flatMap { case (thread, msgs) =>
-        val url = thread.url.getOrElse("")  // needs to change when we have detached threads
-        val msgsWithModifiedAuxData = msgs.map { m =>
-          messagingUtils.modifyMessageWithAuxData(m)
-        }
-        Future.sequence(msgsWithModifiedAuxData).map { completeMsgs =>
-          Ok(Json.obj("id" -> threadId, "uri" -> url, "messages" -> completeMsgs.reverse))
-        }
-      }
+    basicMessageCommander.getThreadMessagesWithBasicUser(ExternalId[MessageThread](threadId), None) map { case (thread, msgs) =>
+      val url = thread.url.getOrElse("")  // needs to change when we have detached threads
+      Ok(Json.obj("id" -> threadId, "uri" -> url, "messages" -> msgs.reverse))
+    }
   }
 
   def getCompactThread(threadId: String) = JsonAction.authenticatedAsync { request =>
-    messagingCommander.getThreadMessagesWithBasicUser(ExternalId[MessageThread](threadId), None) flatMap { case (thread, msgs) =>
+    basicMessageCommander.getThreadMessagesWithBasicUser(ExternalId[MessageThread](threadId), None) map { case (thread, msgs) =>
       val url = thread.url.getOrElse("")  // needs to change when we have detached threads
       val nUrl = thread.nUrl.getOrElse("")  // needs to change when we have detached threads
-      val msgsWithModifiedAuxData = msgs.map { m =>
-        messagingUtils.modifyMessageWithAuxData(m)
-      }
-      Future.sequence(msgsWithModifiedAuxData).map { completeMsgs =>
-        val participants: Set[BasicUserLikeEntity] = completeMsgs.map(_.participants).flatten.toSet
-        Ok(Json.obj(
-          "id" -> threadId,
-          "uri" -> url,
-          "nUrl" -> nUrl,
-          "participants" -> participants,
-          "messages" -> (completeMsgs.reverse map { m =>
-            val adderAndAddedOpt: Option[(String,Seq[String])] = m.auxData match {
-              case Some(JsArray(Seq(JsString("add_participants"), adderBasicUser, addedBasicUsers))) => Some( adderBasicUser.as[BasicUser].externalId.id -> addedBasicUsers.as[Seq[BasicUser]].map(_.externalId.id) )
-              case _ => None
-            }
-            val msgJson = Json.obj(
-              "id" -> m.id.toString,
-              "time" -> m.createdAt.getMillis,
-              "text" -> m.text,
-              "userId" -> m.user.map(_.externalId.toString)
-            )
-            adderAndAddedOpt.map { adderAndAdded =>
-              msgJson.deepMerge(Json.obj("added" -> adderAndAdded._2, "userId" -> adderAndAdded._1))
-            } getOrElse{
-              msgJson
-            }
-          })
-        ))
-      }
+      val participants: Set[BasicUserLikeEntity] = msgs.map(_.participants).flatten.toSet
+      Ok(Json.obj(
+        "id" -> threadId,
+        "uri" -> url,
+        "nUrl" -> nUrl,
+        "participants" -> participants,
+        "messages" -> (msgs.reverse map { m =>
+          val adderAndAddedOpt: Option[(String,Seq[String])] = m.auxData match {
+            case Some(JsArray(Seq(JsString("add_participants"), adderBasicUser, addedBasicUsers))) => Some( adderBasicUser.as[BasicUser].externalId.id -> addedBasicUsers.as[Seq[BasicUser]].map(_.externalId.id) )
+            case _ => None
+          }
+          val msgJson = Json.obj(
+            "id" -> m.id.toString,
+            "time" -> m.createdAt.getMillis,
+            "text" -> m.text,
+            "userId" -> m.user.map(_.externalId.toString)
+          )
+          adderAndAddedOpt.map { adderAndAdded =>
+            msgJson.deepMerge(Json.obj("added" -> adderAndAdded._2, "userId" -> adderAndAdded._1))
+          } getOrElse{
+            msgJson
+          }
+        })
+      ))
     }
   }
