+++ kifi-backend/modules/eliza/app/com/keepit/eliza/controllers/shared/SharedWsMessagingController.scala
@@ -2,7 +2,7 @@ package com.keepit.eliza.controllers.shared
 
 import com.keepit.eliza.model._
 import com.keepit.eliza.controllers._
-import com.keepit.eliza.commanders.{MessagingUtils, NotificationCommander, MessagingCommander}
+import com.keepit.eliza.commanders.{MessageFetchingCommander, NotificationCommander, MessagingCommander}
 import com.keepit.common.db.{ExternalId, State}
 import com.keepit.model.{NotificationCategory, User, ExperimentType}
 import com.keepit.common.controller.{BrowserExtensionController, ActionAuthenticator}
@@ -29,8 +29,8 @@ import com.keepit.common.shutdown.ShutdownCommander
 
 class SharedWsMessagingController @Inject() (
     messagingCommander: MessagingCommander,
+    basicMessageCommander: MessageFetchingCommander,
     notificationCommander: NotificationCommander,
-    messagingUtils: MessagingUtils,
     actionAuthenticator: ActionAuthenticator,
     protected val websocketRouter: WebSocketRouter,
     amazonInstanceInfo: AmazonInstanceInfo,
@@ -71,14 +71,9 @@ class SharedWsMessagingController @Inject() (
     },
     "get_thread" -> { case JsString(threadId) +: _ =>
       log.info(s"[get_thread] user ${socket.userId} thread $threadId")
-      messagingCommander.getThreadMessagesWithBasicUser(ExternalId[MessageThread](threadId), None) map { case (thread, msgs) =>
+      basicMessageCommander.getThreadMessagesWithBasicUser(ExternalId[MessageThread](threadId), None) map { case (thread, msgs) =>
         val url = thread.url.getOrElse("")  // needs to change when we have detached threads
-        val msgsWithModifiedAuxData = msgs.map { m =>
-          messagingUtils.modifyMessageWithAuxData(m)
-        }
-        SafeFuture(Future.sequence(msgsWithModifiedAuxData).map { completeMsgs =>
-          socket.channel.push(Json.arr("thread", Json.obj("id" -> threadId, "uri" -> url, "messages" -> completeMsgs.reverse)))
-        })
+        SafeFuture(socket.channel.push(Json.arr("thread", Json.obj("id" -> threadId, "uri" -> url, "messages" -> msgs.reverse))))
       }
     },
     "add_participants_to_thread" -> { case JsString(threadId) +: JsArray(extUserIds) +: _ =>
